+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Version 0.10 (beta-10)
 
 - Support unspecified address structure
-- It is forbidden to include `. ` in the structure name is prohibited.
+- Forbidden to include `. ` in the structure and the field name are prohibited.
 - bugfix: Crash sometimes occurs when defining a malformed `struct`
 - bugfix: Crash if `}` is specified as the last line without a corresponding `{`.
 - numeric error message detailed